+++ bff/bff-main/src/app.module.ts
@@ -1,0 +1,16 @@
+import { Module } from '@nestjs/common';
+import { ConfigModule } from '@nestjs/config';
+import { ViewsModule } from './views/views.module';
+import { ClientsModule } from './clients/clients.module';
+
+@Module({
+  imports: [
+    ConfigModule.forRoot({
+      isGlobal: true,
+      envFilePath: '.env',
+    }),
+    ClientsModule,
+    ViewsModule,
+  ],
+})
+export class AppModule {}