+++ bff/bff-main/src/views/stocks/stocks.controller.ts
@@ -1,0 +1,17 @@
+import { Controller, Get, Param } from '@nestjs/common';
+import { StocksService } from './stocks.service';
+
+@Controller('views/stocks')
+export class StocksController {
+  constructor(private readonly stocksService: StocksService) {}
+
+  @Get()
+  async getStocks() {
+    return this.stocksService.getAllStocks();
+  }
+
+  @Get(':ticker')
+  async getStock(@Param('ticker') ticker: string) {
+    return this.stocksService.getStock(ticker);
+  }
+}