+++ bff/bff-main/src/views/stocks/stocks.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { ClientsModule } from '../../clients/clients.module';
+import { StocksController } from './stocks.controller';
+import { StocksService } from './stocks.service';
+
+@Module({
+  imports: [ClientsModule],
+  controllers: [StocksController],
+  providers: [StocksService],
+})
+export class StocksModule {}